Updating the parser class
